Refactored adding/deleting forms, added animation toggle
Adding extra scripting around the addition/deletion of forms was
difficult, as there was no simple way of detecting if a form was active
or deleted. Additionally, there was no way of disabling form animations

In this revision:

* deleted forms can be detected using the `data-formset-form-deleted`
attribute. If present, the form is deleted.  If not, it is active.

* `formAdded` events correctly fire on just the form, instead of the
whole empty form HTML snippet.

* Animations for form deletion/addition are now opt-in, instead of
always on. Developers can code their own solution for animations using
the `formAdded`/`formDeleted` events if they need to.