Trigger `formAdded`/`formDeleted` always
The events were not being triggered when there was no delete checkbox,
such as when the Django Formset had `can_delete=False`. Forms can still
be added when `can_delete=False`, so this even should still work.

Additionally, events were being triggered for each form when the Formset
was initialised, but before the calling code could have a chance to
register any events. The initial trigger is now in a `setTimeout()` to
allow the caller to register events.

Fixes #4